+++ src/Dommel/DommelMapper.Delete.cs
@@ -1,0 +1,149 @@
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Data;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
+using Dapper;
+
+namespace Dommel
+{
+    public static partial class DommelMapper
+    {
+        private static readonly ConcurrentDictionary<Type, string> _deleteQueryCache = new ConcurrentDictionary<Type, string>();
+        private static readonly ConcurrentDictionary<Type, string> _deleteAllQueryCache = new ConcurrentDictionary<Type, string>();
+
+        /// <summary>
+        /// Deletes the specified entity from the database.
+        /// Returns a value indicating whether the operation succeeded.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <param name="connection">The connection to the database. This can either be open or closed.</param>
+        /// <param name="entity">The entity to be deleted.</param>
+        /// <param name="transaction">Optional transaction for the command.</param>
+        /// <returns>A value indicating whether the delete operation succeeded.</returns>
+        public static bool Delete<TEntity>(this IDbConnection connection, TEntity entity, IDbTransaction transaction = null)
+        {
+            var sql = BuildDeleteQuery(typeof(TEntity));
+            return connection.Execute(sql, entity, transaction) > 0;
+        }
+
+        /// <summary>
+        /// Deletes the specified entity from the database.
+        /// Returns a value indicating whether the operation succeeded.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <param name="connection">The connection to the database. This can either be open or closed.</param>
+        /// <param name="entity">The entity to be deleted.</param>
+        /// <param name="transaction">Optional transaction for the command.</param>
+        /// <returns>A value indicating whether the delete operation succeeded.</returns>
+        public static async Task<bool> DeleteAsync<TEntity>(this IDbConnection connection, TEntity entity, IDbTransaction transaction = null)
+        {
+            var sql = BuildDeleteQuery(typeof(TEntity));
+            return await connection.ExecuteAsync(sql, entity, transaction) > 0;
+        }
+
+        private static string BuildDeleteQuery(Type type)
+        {
+            if (!_deleteQueryCache.TryGetValue(type, out var sql))
+            {
+                var tableName = Resolvers.Table(type);
+                var keyProperty = Resolvers.KeyProperty(type);
+                var keyColumnName = Resolvers.Column(keyProperty);
+
+                sql = $"delete from {tableName} where {keyColumnName} = @{keyProperty.Name}";
+
+                _deleteQueryCache.TryAdd(type, sql);
+            }
+
+            return sql;
+        }
+
+        /// <summary>
+        /// Deletes all entities of type <typeparamref name="TEntity"/> matching the specified predicate from the database.
+        /// Returns a value indicating whether the operation succeeded.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <param name="connection">The connection to the database. This can either be open or closed.</param>
+        /// <param name="predicate">A predicate to filter which entities are deleted.</param>
+        /// <param name="transaction">Optional transaction for the command.</param>
+        /// <returns>A value indicating whether the delete operation succeeded.</returns>
+        public static bool DeleteMultiple<TEntity>(this IDbConnection connection, Expression<Func<TEntity, bool>> predicate, IDbTransaction transaction = null)
+        {
+            var sql = BuildDeleteMultipleQuery(predicate, out var parameters);
+            return connection.Execute(sql, parameters, transaction) > 0;
+        }
+
+        /// <summary>
+        /// Deletes all entities of type <typeparamref name="TEntity"/> matching the specified predicate from the database.
+        /// Returns a value indicating whether the operation succeeded.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <param name="connection">The connection to the database. This can either be open or closed.</param>
+        /// <param name="predicate">A predicate to filter which entities are deleted.</param>
+        /// <param name="transaction">Optional transaction for the command.</param>
+        /// <returns>A value indicating whether the delete operation succeeded.</returns>
+        public static async Task<bool> DeleteMultipleAsync<TEntity>(this IDbConnection connection, Expression<Func<TEntity, bool>> predicate, IDbTransaction transaction = null)
+        {
+            var sql = BuildDeleteMultipleQuery(predicate, out var parameters);
+            return await connection.ExecuteAsync(sql, parameters, transaction) > 0;
+        }
+
+        private static string BuildDeleteMultipleQuery<TEntity>(Expression<Func<TEntity, bool>> predicate, out DynamicParameters parameters)
+        {
+            var type = typeof(TEntity);
+            if (!_deleteAllQueryCache.TryGetValue(type, out var sql))
+            {
+                var tableName = Resolvers.Table(type);
+                sql = $"delete from {tableName}";
+                _deleteAllQueryCache.TryAdd(type, sql);
+            }
+
+            sql += new SqlExpression<TEntity>()
+                .Where(predicate)
+                .ToSql(out parameters);
+            return sql;
+        }
+
+        /// <summary>
+        /// Deletes all entities of type <typeparamref name="TEntity"/> from the database.
+        /// Returns a value indicating whether the operation succeeded.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <param name="connection">The connection to the database. This can either be open or closed.</param>
+        /// <param name="transaction">Optional transaction for the command.</param>
+        /// <returns>A value indicating whether the delete operation succeeded.</returns>
+        public static bool DeleteAll<TEntity>(this IDbConnection connection, IDbTransaction transaction = null)
+        {
+            var sql = BuildDeleteAllQuery(typeof(TEntity));
+            return connection.Execute(sql, transaction: transaction) > 0;
+        }
+
+        /// <summary>
+        /// Deletes all entities of type <typeparamref name="TEntity"/> from the database.
+        /// Returns a value indicating whether the operation succeeded.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <param name="connection">The connection to the database. This can either be open or closed.</param>
+        /// <param name="transaction">Optional transaction for the command.</param>
+        /// <returns>A value indicating whether the delete operation succeeded.</returns>
+        public static async Task<bool> DeleteAllAsync<TEntity>(this IDbConnection connection, IDbTransaction transaction = null)
+        {
+            var sql = BuildDeleteAllQuery(typeof(TEntity));
+            return await connection.ExecuteAsync(sql, transaction: transaction) > 0;
+        }
+
+        private static string BuildDeleteAllQuery(Type type)
+        {
+            if (!_deleteAllQueryCache.TryGetValue(type, out var sql))
+            {
+                var tableName = Resolvers.Table(type);
+                sql = $"delete from {tableName}";
+                _deleteAllQueryCache.TryAdd(type, sql);
+            }
+
+            return sql;
+        }
+    }
+}
