+++ src/Dommel/DommelMapper.ForeignKeyRelation.cs
@@ -1,0 +1,26 @@
+﻿namespace Dommel
+{
+    public static partial class DommelMapper
+    {
+        /// <summary>
+        /// Describes a foreign key relationship.
+        /// </summary>
+        public enum ForeignKeyRelation
+        {
+            /// <summary>
+            /// Specifies a one-to-one relationship.
+            /// </summary>
+            OneToOne,
+
+            /// <summary>
+            /// Specifies a one-to-many relationship.
+            /// </summary>
+            OneToMany,
+
+            /// <summary>
+            /// Specifies a many-to-many relationship.
+            /// </summary>
+            ManyToMany
+        }
+    }
+}
