+++ src/Dommel/DommelMapper.SqlServerCeSqlBuilder.cs
@@ -1,0 +1,15 @@
+﻿using System.Reflection;
+
+namespace Dommel
+{
+    public static partial class DommelMapper
+    {
+        private sealed class SqlServerCeSqlBuilder : ISqlBuilder
+        {
+            public string BuildInsert(string tableName, string[] columnNames, string[] paramNames, PropertyInfo keyProperty)
+            {
+                return $"insert into {tableName} ({string.Join(", ", columnNames)}) values ({string.Join(", ", paramNames)}) select cast(@@IDENTITY as int)";
+            }
+        }
+    }
+}
