+++ src/Dommel/DommelMapper.SqliteSqlBuilder.cs
@@ -1,0 +1,15 @@
+﻿using System.Reflection;
+
+namespace Dommel
+{
+    public static partial class DommelMapper
+    {
+        private sealed class SqliteSqlBuilder : ISqlBuilder
+        {
+            public string BuildInsert(string tableName, string[] columnNames, string[] paramNames, PropertyInfo keyProperty)
+            {
+                return $"insert into {tableName} ({string.Join(", ", columnNames)}) values ({string.Join(", ", paramNames)}); select last_insert_rowid() id";
+            }
+        }
+    }
+}
